requirements.txt
Flask==2.3.2
numpy==1.23.5
pandas==2.0.2
shap==0.41.0
xgboost==1.7.6
mlflow==2.4.1
mlflow-export-import==1.2.0
cloudpickle==2.2.1
pytest-flask=1.2.0